test(mongodb): consolidate TLS integration coverage
