+++ dt-tests/tests/mongo_to_mongo/mod.rs
@@ -5,5 +5,4 @@ mod review_tests;
 mod revise_tests;
 mod snapshot_tests;
 mod struct_tests;
-mod tls_sharding_tests;
 mod tls_tests;
