+++ dt-tests/tests/mongo_to_mongo/tls/require_sharding_cdc/src_prepare.sql
@@ -1,0 +1,2 @@
+use mongo_tls_sharding_cdc;
+db.dropDatabase();
