+++ dt-tests/tests/mongo_to_mongo/tls/require_sharding_cdc/task_config.ini
@@ -1,0 +1,46 @@
+[extractor]
+db_type=mongo
+extract_type=cdc
+url={mongo_tls_sharding_extractor_url}
+username=
+password=
+ssl_mode=require
+ssl_ca_path={mongo_extractor_ssl_ca_path}
+source=change_stream
+
+[filter]
+ignore_dbs=
+do_dbs=
+do_tbs=mongo_tls_sharding_cdc.*
+ignore_tbs=
+do_events=insert,update,delete
+do_ddls=*
+
+[sinker]
+db_type=mongo
+sink_type=write
+url={mongo_tls_sharding_sinker_url}
+username=
+password=
+ssl_mode=require
+ssl_ca_path={mongo_sinker_ssl_ca_path}
+batch_size=2
+mongo_require_shard_key_filter=true
+
+[router]
+tb_map=
+col_map=
+db_map=
+
+[pipeline]
+buffer_size=4
+checkpoint_interval_secs=1
+
+[parallelizer]
+parallel_type=mongo
+parallel_size=2
+
+[runtime]
+log_dir=./logs
+log_level=info
+log4rs_file=./dt-tests/tests/mongo_to_mongo/tls/log4rs.yaml
