+++ dt-tests/tests/mongo_to_mongo/tls/require_sharding_snapshot/dst_prepare.sql
