+++ dt-tests/tests/mongo_to_mongo/tls/require_sharding_snapshot/src_prepare.sql
