+++ dt-tests/tests/mongo_to_mongo/tls/require_sharding_struct/task_config.ini
@@ -1,0 +1,43 @@
+[extractor]
+db_type=mongo
+extract_type=struct
+url={mongo_tls_sharding_extractor_url}
+username=
+password=
+ssl_mode=require
+ssl_ca_path={mongo_extractor_ssl_ca_path}
+
+[sinker]
+db_type=mongo
+sink_type=struct
+url={mongo_tls_sharding_sinker_url}
+username=
+password=
+ssl_mode=require
+ssl_ca_path={mongo_sinker_ssl_ca_path}
+
+[filter]
+do_dbs=
+ignore_dbs=
+do_tbs=mongo_tls_sharding_struct.*
+ignore_tbs=
+do_events=
+do_structures=collection,shardkey
+
+[router]
+db_map=
+tb_map=
+col_map=
+
+[parallelizer]
+parallel_type=serial
+parallel_size=1
+
+[pipeline]
+buffer_size=4
+checkpoint_interval_secs=1
+
+[runtime]
+log_level=info
+log4rs_file=./dt-tests/tests/mongo_to_mongo/tls/log4rs.yaml
+log_dir=./logs
